--- a/activerecord/lib/active_record/associations/collection_association.rb
+++ b/activerecord/lib/active_record/associations/collection_association.rb
@@ -79,6 +79,7 @@ module ActiveRecord
       def reset
         super
         @target = []
+        @target_index_map = nil
         @replaced_or_added_targets = Set.new
         @association_ids = nil
       end
@@ -252,7 +253,7 @@ module ActiveRecord
           if record.new_record?
             include_in_memory?(record)
           else
-            loaded? ? target.include?(record) : scope.exists?(record.id)
+            loaded? ? !!index_in_target(record) : scope.exists?(record.id)
           end
         else
           false
@@ -262,6 +263,7 @@ module ActiveRecord
       def load_target
         if find_target?
           @target = merge_target_lists(find_target, target)
+          @target_index_map = nil
         end
 
         loaded!
@@ -390,6 +392,7 @@ module ActiveRecord
           delete_records(existing_records, method) if existing_records.any?
           @target -= records
           @association_ids = nil
+          @target_index_map = nil
 
           records.each { |record| callback(:after_remove, record) }
         end
@@ -406,6 +409,7 @@ module ActiveRecord
 
           unless concat(difference(new_target, target))
             @target = original_target
+            @target_index_map = nil
             raise RecordNotSaved, "Failed to replace #{reflection.name} because one or more of the " \
                                   "new records could not be saved."
           end
@@ -442,7 +446,7 @@ module ActiveRecord
 
         def replace_on_target(record, skip_callbacks, replace:, inversing: false)
           if replace && (!record.new_record? || @replaced_or_added_targets.include?(record))
-            index = @target.index(record)
+            index = index_in_target(record)
           end
 
           catch(:abort) do
@@ -456,7 +460,7 @@ module ActiveRecord
           yield(record) if block_given?
 
           if !index && @replaced_or_added_targets.include?(record)
-            index = @target.index(record)
+            index = index_in_target(record)
           end
 
           @replaced_or_added_targets << record if inversing || index || record.new_record?
@@ -465,6 +469,7 @@ module ActiveRecord
             target[index] = record
           elsif @_was_loaded || !loaded?
             @association_ids = nil
+            @target_index_map[record] = target.length if @target_index_map
             target << record
           end
 
@@ -496,9 +501,9 @@ module ActiveRecord
             assoc.reader.any? { |source|
               target_reflection = source.send(reflection.source_reflection.name)
               target_reflection.respond_to?(:include?) ? target_reflection.include?(record) : target_reflection == record
-            } || target.include?(record)
+            } || !!index_in_target(record)
           else
-            target.include?(record)
+            !!index_in_target(record)
           end
         end
 
@@ -515,6 +520,11 @@ module ActiveRecord
           else
             load_target.select { |r| ids.include?(r.id.to_s) }
           end
+        end
+
+        def index_in_target(record)
+          @target_index_map ||= @target.each_with_index.to_h
+          @target_index_map[record]
         end
     end
   end