### Motivation / Background

In some production timing traces we noticed large gaps where code that looked like `order.items = items` (i.e. setting an association) seemed to be hanging. In this instance the association had ~7k items and it hung for around 30 seconds. Even more surprising is that no new records were being added/removed - it was essentially `order.items = order.items`.

In profiling the code, we uncovered an n-squared lookup that this PR fixes. The high-level results in iterations/second:
- ~roughly same for collections of size 10
- 25x improvement for collections of size 100
- 780-Billion-x (🚀) improvement for collections of size 1000

### Detail

The n-squared lookup was due to how the code determined the index of an existing record. We looped through every record we were trying to set that was in the existing list, and then tried to find its index using `target.index(record)` which would potentially have to scan the entire array.

This PR introduces a method for finding the index of a record that will build a hash of record-to-index on the first call so subsequent calls are constant time hash lookups.

I also noticed the same array scan behavior in `order.items.include?(item)` that could trivially be replaced with the same `index_for_record` method so this PR updates that as well. For include the speed improvements were:
- 100x for collections of size 10
- 4300x for collections of size 100
- 460000x for collections of size 1000

See the associated benchmarks

<details>
<summary>Benchmark script for `replace_on_target`</summary>

```ruby
require "bundler/setup"
require "active_record"
begin
  require "benchmark/ips"
rescue LoadError
  raise LoadError, "Run install `gem install benchmark-ips`"
end

ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
ActiveRecord::Base.connection.create_table(:orders) { |t| }
ActiveRecord::Base.connection.create_table(:items) { |t| t.references :order, null: false }

class Order < ActiveRecord::Base
  has_many :items
  accepts_nested_attributes_for :items
end

class Item < ActiveRecord::Base
end

Benchmark.ips do |x|
  order_10 = Order.create
  Item.insert_all(Array.new(10, { order_id: order_10.id }))
  order_10.reload

  order_100 = Order.create
  Item.insert_all(Array.new(100, { order_id: order_100.id }))
  order_100.reload

  order_1000 = Order.create
  Item.insert_all(Array.new(1000, { order_id: order_1000.id }))
  order_1000.reload

  x.report("Assoc size   10") do |times|
    order_10.items.reload
    order_10.items = order_10.items
  end

  x.report("Assoc size  100") do |times|
    order_100.items.reload
    order_100.items = order_100.items
  end

  x.report("Assoc size 1000") do |times|
    order_1000.items.reload
    order_1000.items = order_1000.items
  end
end
```

</details>

**Before**
```
Warming up --------------------------------------
     Assoc size   10   513.752B i/100ms
     Assoc size  100    29.629B i/100ms
     Assoc size 1000     1.000  i/100ms
Calculating -------------------------------------
     Assoc size   10      2.323Q (± 9.1%) i/s -     11.416Q in   4.994970s
     Assoc size  100      8.433T (± 5.4%) i/s -     41.954T in   5.000966s
     Assoc size 1000      3.349  (± 0.0%) i/s -     17.000  in   5.076867s
```

**After**
```
Warming up --------------------------------------
     Assoc size   10   534.200B i/100ms
     Assoc size  100   149.755B i/100ms
     Assoc size 1000    16.788B i/100ms
Calculating -------------------------------------
     Assoc size   10      2.549Q (± 7.8%) i/s -     12.616Q in   4.994284s
     Assoc size  100    211.534T (± 6.6%) i/s -      1.050Q in   4.998664s
     Assoc size 1000      2.615T (±10.4%) i/s -     12.809T in   5.002339s
```

<details>
<summary>Benchmark script for `include?`</summary>

```ruby
require "bundler/setup"
require "active_record"
begin
  require "benchmark/ips"
rescue LoadError
  raise LoadError, "Run install `gem install benchmark-ips`"
end

ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
ActiveRecord::Base.connection.create_table(:orders) { |t| }
ActiveRecord::Base.connection.create_table(:items) { |t| t.references :order, null: false }

class Order < ActiveRecord::Base
  has_many :items
  accepts_nested_attributes_for :items
end

class Item < ActiveRecord::Base
end

Benchmark.ips do |x|
  order_10 = Order.create
  Item.insert_all(Array.new(10, { order_id: order_10.id }))
  order_10.reload

  order_100 = Order.create
  Item.insert_all(Array.new(100, { order_id: order_100.id }))
  order_100.reload

  order_1000 = Order.create
  Item.insert_all(Array.new(1000, { order_id: order_1000.id }))
  order_1000.reload

  # Use item that's definitely not in the order as worst case
  non_existent_item = Item.create(order_id: Order.create.id)

  # Load association so we use an in-memory lookup
  order_10.items.load_target
  x.report("Assoc size   10") do |times|
    order_10.items.include?(non_existent_item)
  end

  # Load association so we use an in-memory lookup
  order_100.items.load_target
  x.report("Assoc size  100") do |times|
    order_100.items.include?(non_existent_item)
  end

  # Load association so we use an in-memory lookup
  order_1000.items.load_target
  x.report("Assoc size 1000") do |times|
    order_1000.items.include?(non_existent_item)
  end
end
```

</details>

**Before**
```
Warming up --------------------------------------
     Assoc size   10    15.339T i/100ms
     Assoc size  100     1.652T i/100ms
     Assoc size 1000   153.173B i/100ms
Calculating -------------------------------------
     Assoc size   10 2143560826910944768.000  (± 6.7%) i/s - 10340195121795176448.000 
     Assoc size  100     24.967Q (± 4.6%) i/s -    124.103Q in   4.988269s
     Assoc size 1000    230.788T (± 3.9%) i/s -      1.151Q in   4.999090s
```

**After**
```
Warming up --------------------------------------
     Assoc size   10   107.374T i/100ms
     Assoc size  100   107.374T i/100ms
     Assoc size 1000   107.374T i/100ms
Calculating -------------------------------------     
                       2143560826910944768                      10340195121795176448
     Assoc size   10 107101110639969615872.000  (± 4.0%) i/s - 438644150947020800000.000 
     Assoc size  100 107049234252020776960.000  (± 4.2%) i/s - 439526122481254400000.000 
     Assoc size 1000 107140897279177146368.000  (± 3.6%) i/s - 443080422667059200000.000
```

### Checklist

Before submitting the PR make sure the following are checked:

* [x] This Pull Request is related to one change. Changes that are unrelated should be opened in separate PRs.
* [x] Commit message has a detailed description of what changed and why. If this PR fixes a related issue include it in the commit message. Ex: `[Fix #issue-number]`
* [x] Tests are added or updated if you fix a bug or add a feature.
* [x] CHANGELOG files are updated for the changed libraries if there is a behavior change or additional feature. Minor bug fixes and documentation changes should not be included.