@ghiculescu 2022-12-06
This reminds me of a similar regression we saw when upgrading to 6.1. Just sharing in case that fix is helpful: https://github.com/rails/rails/pull/42524