@jeffcarbs 2022-12-08
I squashed the commits into one (including the suggested change) and just pushed it up. I did make one other change, which was for the usages of `index_in_target` inside the `include?` methods I used `!!` to ensure it returns a boolean.