@jonathanhefner 2022-12-08
> * 100x for collections of size 10
> * 4300x for collections of size 100
> * 460000x for collections of size 1000

These numbers are inaccurate due to the `times` block parameter in your benchmark scripts, e.g.:

> ```ruby
>   x.report("Assoc size 1000") do |times|
>     order_1000.items.reload
>     order_1000.items = order_1000.items
>   end
> ```

When the benchmark block accepts an argument, `benchmark-ips` [expects the block to execute the code the given number of times](https://github.com/evanphx/benchmark-ips/blob/ea645f694667ff24956ddd14727b024bd4f507a0/README.md?plain=1#L39-L43).

Would you mind re-running the benchmarks without the block parameters?

Also, does your first benchmark represent a real world use case?  I had been trying to refactor some of this code in #46652.  I haven't merged that PR yet (because it caused a failure in Shopify's test suite that has yet to be investigated), but it has benchmark script based on the performance concerns from #42524.  Would you mind trying that benchmark with your code?  (You may need to add a few `require` statements; I ran it using `bin/rails r` with a fresh app pointing at the appropriate Rails branch.)