@jeffcarbs 2022-12-08
@jonathanhefner - Thanks for the reply and pointing out the issues with the benchmarks. I'll update my code and post the new benchmarks.

Also, in your comment I think you may have referenced the wrong PR? You said:
> I had been trying to refactor some of this code in #46652

which is this PR. Did you mean to link #45399 by chance or maybe a different one? Or maybe I misunderstood the comment.

> Also, does your first benchmark represent a real world use case

Do you mean does `order.items = items` on a huge collection represent a real-world use-case? If so, the answer is yes for us. We have a job that runs to ensure the correct "bars" are associated with the right "foo", so it gathers up the correct set of "bars" and does `foo.bars = bars` and in production this collection can be upwards of 10k things. In most cases, the result will just be a no-op or adding/removing a few records since the process runs frequently as the "bars" are modified.

Given this performance issue we uncovered, we've refactored that code to manually create/remove the join records since we can use bulk operations like `insert_all` and `delete_all` (would love if rails had a built-in way to bulk update join records 🙂)