@jeffcarbs 2022-12-08
@jonathanhefner - I re-ran the benchmarks and updated the PR description with the correct scripts and results. Thanks again for pointing out the mistake with the block variable, billion-x improvements were definitely suspicious 😅.

Speed improvements for `replace_on_target`:
- Size 10: roughly same
- Size 100: ~5x improvement
- Size 1000: ~50x improvement

Speed improvements for replace_on_target:
- Size 10: ~7x
- Size 100: ~68x improvement
- Size 1000: ~690x improvement

I tried running the benchmarks from #45399, here are my results:

Results with `main` branch:
```
~ [main] $ ruby benchmark_from_45399.rb
-- create_table(:authors, {:force=>true})
   -> 0.0103s
-- create_table(:posts, {:force=>true})
   -> 0.0002s
has_many_inversing == false, posts.count == 1     => `posts.build` @ 0.043ms
has_many_inversing == false, posts.count == 10    => `posts.build` @ 0.043ms
has_many_inversing == false, posts.count == 100   => `posts.build` @ 0.046ms
has_many_inversing == false, posts.count == 1000  => `posts.build` @ 0.041ms
has_many_inversing == true, posts.count == 1      => `posts.build` @ 0.041ms
has_many_inversing == true, posts.count == 10     => `posts.build` @ 0.040ms
has_many_inversing == true, posts.count == 100    => `posts.build` @ 0.039ms
has_many_inversing == true, posts.count == 1000   => `posts.build` @ 0.041ms
has_many_inversing == false, posts.count == 1     => `posts << post` @ 0.063ms
has_many_inversing == false, posts.count == 10    => `posts << post` @ 0.066ms
has_many_inversing == false, posts.count == 100   => `posts << post` @ 0.066ms
has_many_inversing == false, posts.count == 1000  => `posts << post` @ 0.068ms
has_many_inversing == true, posts.count == 1      => `posts << post` @ 0.066ms
has_many_inversing == true, posts.count == 10     => `posts << post` @ 0.067ms
has_many_inversing == true, posts.count == 100    => `posts << post` @ 0.067ms
has_many_inversing == true, posts.count == 1000   => `posts << post` @ 0.068ms
```

Results with this branch:
```
~ [collection-association-performance %] $ ruby benchmark_from_45399.rb
-- create_table(:authors, {:force=>true})
   -> 0.0112s
-- create_table(:posts, {:force=>true})
   -> 0.0003s
has_many_inversing == false, posts.count == 1     => `posts.build` @ 0.043ms
has_many_inversing == false, posts.count == 10    => `posts.build` @ 0.044ms
has_many_inversing == false, posts.count == 100   => `posts.build` @ 0.045ms
has_many_inversing == false, posts.count == 1000  => `posts.build` @ 0.041ms
has_many_inversing == true, posts.count == 1      => `posts.build` @ 0.038ms
has_many_inversing == true, posts.count == 10     => `posts.build` @ 0.040ms
has_many_inversing == true, posts.count == 100    => `posts.build` @ 0.040ms
has_many_inversing == true, posts.count == 1000   => `posts.build` @ 0.042ms
has_many_inversing == false, posts.count == 1     => `posts << post` @ 0.065ms
has_many_inversing == false, posts.count == 10    => `posts << post` @ 0.067ms
has_many_inversing == false, posts.count == 100   => `posts << post` @ 0.064ms
has_many_inversing == false, posts.count == 1000  => `posts << post` @ 0.066ms
has_many_inversing == true, posts.count == 1      => `posts << post` @ 0.063ms
has_many_inversing == true, posts.count == 10     => `posts << post` @ 0.067ms
has_many_inversing == true, posts.count == 100    => `posts << post` @ 0.064ms
has_many_inversing == true, posts.count == 1000   => `posts << post` @ 0.070ms
```

Also, just to have more data and be able to more easily compare, I pulled down the `jonathanhefner:collection_association-refactor-inversing-logic` branch and ran all of the benchmarks (performance similar to `main`):
```
~[(77889e87c0...) *%] $ ruby benchmark_replace_on_target.rb
Warming up --------------------------------------
     Assoc size   10   465.000  i/100ms
     Assoc size  100    28.000  i/100ms
     Assoc size 1000     1.000  i/100ms
Calculating -------------------------------------
     Assoc size   10      4.702k (± 1.5%) i/s -     23.715k in   5.045185s
     Assoc size  100    287.962  (± 0.7%) i/s -      1.456k in   5.056606s
     Assoc size 1000      3.324  (± 0.0%) i/s -     17.000  in   5.116352s

~[(77889e87c0...) *%] $ uby benchmark_include.rb
Warming up --------------------------------------
     Assoc size   10    14.127k i/100ms
     Assoc size  100     1.531k i/100ms
     Assoc size 1000   152.000  i/100ms
Calculating -------------------------------------
     Assoc size   10    142.562k (± 1.0%) i/s -    720.477k in   5.054247s
     Assoc size  100     15.351k (± 0.8%) i/s -     78.081k in   5.086709s
     Assoc size 1000      1.529k (± 1.2%) i/s -      7.752k in   5.070870s

~[(77889e87c0...) *%] $ ruby benchmark_from_45399.rb
-- create_table(:authors, {:force=>true})
   -> 0.0106s
-- create_table(:posts, {:force=>true})
   -> 0.0002s
has_many_inversing == false, posts.count == 1     => `posts.build` @ 0.039ms
has_many_inversing == false, posts.count == 10    => `posts.build` @ 0.038ms
has_many_inversing == false, posts.count == 100   => `posts.build` @ 0.039ms
has_many_inversing == false, posts.count == 1000  => `posts.build` @ 0.042ms
has_many_inversing == true, posts.count == 1      => `posts.build` @ 0.038ms
has_many_inversing == true, posts.count == 10     => `posts.build` @ 0.035ms
has_many_inversing == true, posts.count == 100    => `posts.build` @ 0.036ms
has_many_inversing == true, posts.count == 1000   => `posts.build` @ 0.043ms
has_many_inversing == false, posts.count == 1     => `posts << post` @ 0.067ms
has_many_inversing == false, posts.count == 10    => `posts << post` @ 0.065ms
has_many_inversing == false, posts.count == 100   => `posts << post` @ 0.065ms
has_many_inversing == false, posts.count == 1000  => `posts << post` @ 0.070ms
has_many_inversing == true, posts.count == 1      => `posts << post` @ 0.065ms
has_many_inversing == true, posts.count == 10     => `posts << post` @ 0.068ms
has_many_inversing == true, posts.count == 100    => `posts << post` @ 0.066ms
has_many_inversing == true, posts.count == 1000   => `posts << post` @ 0.070ms
```